requirements.txt
flask
tensorflow
pillow
gunicorn
gdown